Finish the split
