+++ lib/logger.js
@@ -1,5 +1,5 @@
 import { logger } from 'appium-support';
 
-const log = logger.getLogger('Simctl');
+const log = logger.getLogger('simctl');
 
 export default log;
