+++ lib/subcommands/boot.js
@@ -13,7 +13,7 @@ const commands = {};
 commands.bootDevice = async function bootDevice () {
   try {
     await this.exec('boot', {
-      args: [this._requireUdid('boot')]
+      args: [this.requireUdid('boot')]
     });
   } catch (e) {
     if (_.includes(e.message, 'Unable to boot device in current state: Booted')) {
