+++ lib/subcommands/install.js
@@ -11,7 +11,7 @@ const commands = {};
  */
 commands.installApp = async function installApp (appPath) {
   await this.exec('install', {
-    args: [this._requireUdid('install'), appPath],
+    args: [this.requireUdid('install'), appPath],
   });
 };
