+++ lib/subcommands/terminate.js
@@ -1,0 +1,18 @@
+const commands = {};
+
+/**
+ * Terminate the given running application on Simulator.
+ * It is required that Simulator is in _booted_ state.
+ *
+ * @param {string} bundleId - Bundle identifier of the application,
+ *                            which is going to be terminated.
+ * @throws {Error} If the corresponding simctl subcommand command
+ *                 returns non-zero return code.
+ */
+commands.terminateApp = async function terminateApp (bundleId) {
+  await this.exec('terminate', {
+    args: [this.requireUdid('terminate'), bundleId],
+  });
+};
+
+export default commands;
